+++ CHANGES.md
@@ -6,7 +6,7 @@ Here you can see the full list of changes between each pytest-easyread release.
 
 - Add pathname to failure reporting output
 - Change pytest backwards compatibility info to be more accurate
-
+- Test that pytest-easyread works well with parametrize
 
 ## 0.0.0 (August 11, 2017)